requirements.txt
flask==3.0.0
requests==2.31.0
selenium==4.15.2
openai==1.54.3
python-dotenv==1.0.0
gunicorn==21.2.0
pandas==2.1.3
numpy==1.26.2
Pillow==10.1.0
pdfplumber==0.10.3
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0